pyproject.toml
[tool.poetry]
name = "telegram-proc-control"
version = "0.1.0"
description = ""
authors = ["Nicolas Chan <nicolas@nicolaschan.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
